Cleaning up base connection.
